# cmd/api/api.swagger.yaml
openapi: 3.0.3
info:
  title: Swagger для проекта CourseHub
  description: Основная документация по Swagger.
  contact:
    url: https://t.me/CringeINC
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
externalDocs:
  description: chi documentation =)
  url: https://go-chi.io/#/
tags:
  - name: users
    description: Список операций с пользователями
  - name: posts
    description: Список постов пользователя
  - name: groups
    description: Список групп пользователей
  - name: health
    description: pong
paths:
  /pong:
    get:
      tags:
        - health
      summary: pong
      operationId: pong
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: string
  /posts:
    get:
      tags:
        - posts
      summary: Get all posts
      operationId: getPosts
      parameters:
        - name: limit
          in: query
          description: "post limit params"
          required: false
          schema:
            type: integer
            example: "limit=10"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Posts"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServiceError"

  /posts/{id}:
    get:
      tags:
        - posts
      summary: Get post
      operationId: getPost
      parameters:
        - name: id
          in: path
          description: "post id params"
          required: true
          schema:
            type: integer
            example: 1
        - name: title
          in: query
          description: "post title params"
          required: false
          schema:
            type: string
            example: "title=test"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServiceError"
    put:
      tags:
        - posts
      summary: Update post
      operationId: updatePost
      parameters:
        - name: id
          in: path
          description: "post id params"
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServiceError"
    delete:
      tags:
        - posts
      summary: Delete post
      operationId: deletePost
      parameters:
        - name: id
          in: path
          description: "post id params"
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServiceError"
  /posts/like:
    get:
      tags:
        - posts
      summary: Get Likes for post
      operationId: getLikes
      parameters:
        - name: post_id
          in: query
          description: "post id params"
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServiceError"
    post:
      tags:
        - posts
      summary: Like post
      operationId: likePost
      parameters:
        - name: post_id
          in: query
          description: "post id params"
          required: true
          schema:
            type: integer
            example: 1
        - name: user_id
          in: query
          description: "user id params"
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: Successful operation
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServiceError"
  /posts/new:
    post:
      tags:
        - posts
      summary: Create new post
      operationId: createPost
      requestBody:
        description: "Create new post"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Post"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServiceError"

components:
  schemas:
    InternalServiceError:
      type: object
      properties:
        message:
          example: Internal server error
        code:
          type: integer
          example: Internal server error
    Post:
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          example: Java is the best
        body:
          type: string
          example: Java is the best
        images_list:
          type: array
          items:
            type: string
            format: binary
        author_id:
          type: integer
          example: 12
        created_at:
          type: string
          example: "2018-01-01"
        updated_at:
          type: string
          example: "2018-01-01"
    Posts:
      type: array
      items:
        $ref: "#/components/schemas/Post"